add support for float in AcWait

+++ classes/ac_wait.py
@@ -3,8 +3,8 @@ from time import sleep
 
 
 class AcWait(ActionUI):
-    def __init__(self, wait_seconds: int) -> None:
-        self.wait = int(wait_seconds)
+    def __init__(self, wait_seconds: float) -> None:
+        self.wait = float(wait_seconds)
         super().__init__(name="Wait")
 
     def __str__(self) -> str:
